[#78] FinishConfirm으로 앨범 저장 로직 이동
